@@ -15,7 +15,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.8+9
+version: 1.0.9+10
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -38,7 +38,7 @@ dependencies:
   http: 0.13.4
   shared_preferences: 2.0.7
   connectivity: 3.0.6
-  google_maps_flutter: 2.1.1
+  google_maps_flutter: 2.2.5
   permission_handler: 8.3.0
   package_info: ^2.0.2
   url_launcher: 6.0.3
