@@ -20,8 +20,8 @@ BusStopDetail.fromJson(Map<String, dynamic> json):
       englishName = json["name_en"] ?? "",
       TCName = json["name_tc"] ?? "",
       SCName = json["name_sc"] ?? "",
-      latitude = double.tryParse(json["lat"]?? 0) ?? 0,
-      longitude = double.tryParse(json["long"]?? 0) ?? 0,
+      latitude = double.tryParse(json["lat"]?? 0) ?? (json["lat"] is double ? json["lat"] : 0),
+      longitude = double.tryParse(json["long"]?? 0) ?? (json["lat"] is double ? json["long"] : 0),
       timestamp = DateTime.tryParse(json["data_timestamp"]??"") ?? DateTime.now(){
       Map<String, String> nameData = Map();
       nameData["en"] = englishName;
