@@ -9,22 +9,25 @@ import 'package:minimal_bus_hk/utils/stores.dart';
 import 'package:minimal_bus_hk/model/bus_route.dart';
 
 import 'package:mobx/mobx.dart';
+
+import '../utils/localization_util.dart';
+import '../utils/network_util.dart';
+
 part 'route_list_store.g.dart';
 
 class RouteListStore = RouteListStoreBase with _$RouteListStore;
 
 abstract class RouteListStoreBase with Store {
-
   @observable
   String filterKeyword = "";
 
   @action
-  void setFilterKeyword(String keyword){
+  void setFilterKeyword(String keyword) {
     filterKeyword = keyword;
   }
 
   @computed
-  List<String> get _keywords{
+  List<String> get _keywords {
     return filterKeyword.split(" ");
   }
 
@@ -32,10 +35,10 @@ abstract class RouteListStoreBase with Store {
   DirectionalRoute? selectedDirectionalRoute;
 
   @action
-  void setSelectedDirectionalRoute(DirectionalRoute? busRoute){
-    if(selectedDirectionalRoute != busRoute) {
+  void setSelectedDirectionalRoute(DirectionalRoute? busRoute) {
+    if (selectedDirectionalRoute != busRoute) {
       selectedDirectionalRoute = busRoute;
-    }else{
+    } else {
       selectedDirectionalRoute = null;
     }
   }
@@ -44,135 +47,79 @@ abstract class RouteListStoreBase with Store {
   String filterStopIdentifier = "";
 
   @action
-  void setFilterStopIdentifier(String identifier){
+  void setFilterStopIdentifier(String identifier) {
     filterStopIdentifier = identifier;
   }
 
   @action
-  void clearFilters(){
+  void clearFilters() {
     filterKeyword = "";
     filterStopIdentifier = "";
   }
 
-  /*
-  @computed
-  ObservableList<DirectionalRoute> get displayedDirectionalRoutes {
-    if(filterKeyword != null && Stores.dataManager.directionalRouteList != null) {
-
-      var directionalRouteList = filterStopIdentifier != null && filterStopIdentifier.length > 0
-          && Stores.dataManager.stopRoutesMap.containsKey(filterStopIdentifier)?
-      Stores.dataManager.directionalRouteList.where((element) {
-        var directionalRoutesOfStop = Stores.dataManager.stopRoutesMap[filterStopIdentifier];
-        for(DirectionalRoute directionalRoute in directionalRoutesOfStop){
-          if(directionalRoute.route.routeCode == element.route.routeCode && directionalRoute.isInbound == element.isInbound){
-            return true;
-          }
-        }
-        return false;
-      }
-      ) : Stores.dataManager.directionalRouteList;
-
-      var result =  ObservableList<DirectionalRoute>();
-      result.addAll(directionalRouteList.where((element) {
-        if(element == null){
-          return false;
-        }
-        for(var keyword in _keywords) {
-          if(element.route.routeCode.toLowerCase().contains(keyword.toLowerCase())
-            || (element.isInbound &&
-                  ((keyword.length > 1 && element.route.originEnglishName.toLowerCase().contains(
-                  keyword.toLowerCase()))
-                  || element.route.originTCName.toLowerCase().contains(
-                  keyword.toLowerCase())
-                  || element.route.originSCName.toLowerCase().contains(
-                  keyword.toLowerCase())))
-              ||(!element.isInbound &&
-                  ((keyword.length > 1 &&  element.route.destinationEnglishName.toLowerCase().contains(
-                  keyword.toLowerCase()))
-                  || element.route.destinationSCName.toLowerCase().contains(
-                  keyword.toLowerCase())
-                  || element.route.destinationTCName.toLowerCase().contains(
-                  keyword.toLowerCase())))){
-
-          }else{
-            return false;
-          }
-        }
-        return true;
-      }
-      ).toList());
-
-      result.sort((a,b){
-        var result = a.route.routeCode.length.compareTo(b.route.routeCode.length);
-        if(result == 0){
-          result = a.route.routeCode.compareTo(b.route.routeCode);
-        }
-        return result;
-      });
-      return result;
-    }else {
-      var result = Stores.dataManager.directionalRouteList;
-
-      return result;
-    }
-  }
-*/
   @computed
   ObservableList<DirectionalRoute> get displayedDirectionalRoutesForNWFBAndCTB {
-    if(filterKeyword != null && Stores.dataManager.directionalRouteListOfNWFBAndCTB != null) {
+    if (filterKeyword.isNotEmpty &&
+        Stores.dataManager.directionalRouteListOfNWFBAndCTB != null) {
+      var directionalRouteList = filterStopIdentifier != null &&
+              filterStopIdentifier.length > 0 &&
+              Stores.dataManager.stopRoutesMap.containsKey(filterStopIdentifier)
+          ? Stores.dataManager.directionalRouteListOfNWFBAndCTB
+              .where((element) {
+              var directionalRoutesOfStop =
+                  Stores.dataManager.stopRoutesMap[filterStopIdentifier];
+              for (DirectionalRoute directionalRoute
+                  in directionalRoutesOfStop!) {
+                if (directionalRoute.route.routeUniqueIdentifier ==
+                        element.route.routeUniqueIdentifier &&
+                    directionalRoute.isInbound == element.isInbound) {
+                  return true;
+                }
+              }
+              return false;
+            })
+          : Stores.dataManager.directionalRouteListOfNWFBAndCTB;
 
-      var directionalRouteList = filterStopIdentifier != null && filterStopIdentifier.length > 0
-          && Stores.dataManager.stopRoutesMap.containsKey(filterStopIdentifier)?
-      Stores.dataManager.directionalRouteListOfNWFBAndCTB.where((element) {
-        var directionalRoutesOfStop = Stores.dataManager.stopRoutesMap[filterStopIdentifier];
-        for(DirectionalRoute directionalRoute in directionalRoutesOfStop!){
-          if(directionalRoute.route.routeUniqueIdentifier == element.route.routeUniqueIdentifier && directionalRoute.isInbound == element.isInbound){
-            return true;
-          }
-        }
-        return false;
-      }
-      ) : Stores.dataManager.directionalRouteListOfNWFBAndCTB;
-
-      var result =  ObservableList<DirectionalRoute>();
+      var result = ObservableList<DirectionalRoute>();
       result.addAll(directionalRouteList.where((element) {
-        if(element == null){
+        if (element == null) {
           return false;
         }
-        for(var keyword in _keywords) {
-          if(element.route.routeCode.toLowerCase().contains(keyword.toLowerCase())
-              || (element.isInbound &&
-                  ((keyword.length > 1 && element.route.originEnglishName.toLowerCase().contains(
-                      keyword.toLowerCase()))
-                      || element.route.originTCName.contains(
-                          keyword)
-                      || element.route.originSCName.contains(
-                          keyword)))
-              ||(!element.isInbound &&
-                  ((keyword.length > 1 &&  element.route.destinationEnglishName.toLowerCase().contains(
-                      keyword.toLowerCase()))
-                      || element.route.destinationSCName.contains(
-                          keyword)
-                      || element.route.destinationTCName.contains(
-                          keyword)))){
-
-          }else{
+        for (var keyword in _keywords) {
+          if (element.route.routeCode
+                  .toLowerCase()
+                  .contains(keyword.toLowerCase()) ||
+              (element.isInbound &&
+                  ((keyword.isNotEmpty &&
+                          element.route.originEnglishName
+                              .toLowerCase()
+                              .contains(keyword.toLowerCase())) ||
+                      element.route.originTCName.contains(keyword) ||
+                      element.route.originSCName.contains(keyword))) ||
+              (!element.isInbound &&
+                  ((keyword.isNotEmpty &&
+                          element.route.destinationEnglishName
+                              .toLowerCase()
+                              .contains(keyword.toLowerCase())) ||
+                      element.route.destinationSCName.contains(keyword) ||
+                      element.route.destinationTCName.contains(keyword)))) {
+          } else {
             return false;
           }
         }
         return true;
-      }
-      ).toList());
+      }).toList());
 
-      result.sort((a,b){
-        var result = a.route.routeCode.length.compareTo(b.route.routeCode.length);
-        if(result == 0){
+      result.sort((a, b) {
+        var result =
+            a.route.routeCode.length.compareTo(b.route.routeCode.length);
+        if (result == 0) {
           result = a.route.routeCode.compareTo(b.route.routeCode);
         }
         return result;
       });
       return result;
-    }else {
+    } else {
       var result = Stores.dataManager.directionalRouteListOfNWFBAndCTB;
 
       return result;
@@ -181,60 +128,74 @@ abstract class RouteListStoreBase with Store {
 
   @computed
   ObservableList<DirectionalRoute> get displayedDirectionalRoutesForKMB {
-    if(filterKeyword != null && Stores.dataManager.directionalRouteListOfKMB != null) {
+    if (filterKeyword.isNotEmpty &&
+        Stores.dataManager.directionalRouteListOfKMB != null) {
+      var directionalRouteList = filterStopIdentifier != null &&
+              filterStopIdentifier.length > 0 &&
+              Stores.dataManager.stopRoutesMap.containsKey(filterStopIdentifier)
+          ? Stores.dataManager.directionalRouteListOfKMB.where((element) {
+              var directionalRoutesOfStop =
+                  Stores.dataManager.stopRoutesMap[filterStopIdentifier];
+              for (DirectionalRoute directionalRoute
+                  in directionalRoutesOfStop!) {
+                if (directionalRoute.route.routeUniqueIdentifier ==
+                        element.route.routeUniqueIdentifier &&
+                    directionalRoute.isInbound == element.isInbound) {
+                  return true;
+                }
+              }
+              return false;
+            })
+          : Stores.dataManager.directionalRouteListOfKMB;
 
-      var directionalRouteList = filterStopIdentifier != null && filterStopIdentifier.length > 0
-          && Stores.dataManager.stopRoutesMap.containsKey(filterStopIdentifier)?
-      Stores.dataManager.directionalRouteListOfKMB.where((element) {
-        var directionalRoutesOfStop = Stores.dataManager.stopRoutesMap[filterStopIdentifier];
-        for(DirectionalRoute directionalRoute in directionalRoutesOfStop!){
-          if(directionalRoute.route.routeUniqueIdentifier == element.route.routeUniqueIdentifier && directionalRoute.isInbound == element.isInbound){
-            return true;
-          }
-        }
-        return false;
-      }
-      ) : Stores.dataManager.directionalRouteListOfKMB;
-
-      var result =  ObservableList<DirectionalRoute>();
+      var result = ObservableList<DirectionalRoute>();
       result.addAll(directionalRouteList.where((element) {
-        if(element == null){
+        if (element == null) {
           return false;
         }
-        for(var keyword in _keywords) {
-          if(element.route.routeCode.toLowerCase().contains(keyword.toLowerCase())
-              || (element.isInbound &&
-                  ((keyword.length > 1 && element.route.originEnglishName.toLowerCase().contains(
-                      keyword.toLowerCase()))
-                      || element.route.originTCName.toLowerCase().contains(
-                          keyword.toLowerCase())
-                      || element.route.originSCName.toLowerCase().contains(
-                          keyword.toLowerCase())))
-              ||(!element.isInbound &&
-                  ((keyword.length > 1 &&  element.route.destinationEnglishName.toLowerCase().contains(
-                      keyword.toLowerCase()))
-                      || element.route.destinationSCName.toLowerCase().contains(
-                          keyword.toLowerCase())
-                      || element.route.destinationTCName.toLowerCase().contains(
-                          keyword.toLowerCase())))){
-
-          }else{
+        for (var keyword in _keywords) {
+          if (element.route.routeCode
+                  .toLowerCase()
+                  .contains(keyword.toLowerCase()) ||
+              (element.isInbound &&
+                  ((keyword.isNotEmpty &&
+                          element.route.originEnglishName
+                              .toLowerCase()
+                              .contains(keyword.toLowerCase())) ||
+                      element.route.originTCName
+                          .toLowerCase()
+                          .contains(keyword.toLowerCase()) ||
+                      element.route.originSCName
+                          .toLowerCase()
+                          .contains(keyword.toLowerCase()))) ||
+              (!element.isInbound &&
+                  ((keyword.isNotEmpty &&
+                          element.route.destinationEnglishName
+                              .toLowerCase()
+                              .contains(keyword.toLowerCase())) ||
+                      element.route.destinationSCName
+                          .toLowerCase()
+                          .contains(keyword.toLowerCase()) ||
+                      element.route.destinationTCName
+                          .toLowerCase()
+                          .contains(keyword.toLowerCase())))) {
+          } else {
             return false;
           }
         }
         return true;
-      }
-      ).toList());
+      }).toList());
 
-      result.sort((a,b){
-        var result = a.route.routeCode.length.compareTo(b.route.routeCode.length);
-        if(result == 0){
+      result.sort((a, b) {
+        var result =
+            a.route.routeCode.length.compareTo(b.route.routeCode.length);
+        if (result == 0) {
           result = a.route.routeCode.compareTo(b.route.routeCode);
         }
         return result;
       });
       return result;
-    }else {
+    } else {
       var result = Stores.dataManager.directionalRouteListOfKMB;
 
       return result;
@@ -245,7 +206,31 @@ abstract class RouteListStoreBase with Store {
   bool dataFetchingError = false;
 
   @action
-  void setDataFetchingError(bool hasError){
+  void setDataFetchingError(bool hasError) {
     dataFetchingError = hasError;
+  }
+
+  @computed
+  String get nwfbTabTitle {
+    String title =
+        "${LocalizationUtil.localizedString(NetworkUtil.companyCodeNWFB, Stores.localizationStore.localizationPref)} & ${LocalizationUtil.localizedString(NetworkUtil.companyCodeCTB, Stores.localizationStore.localizationPref)}";
+
+    if (filterKeyword.isNotEmpty) {
+      title = "${title} (${displayedDirectionalRoutesForNWFBAndCTB.length})";
+    }
+
+    return title;
+  }
+
+  @computed
+  String get kmbTabTitle {
+    String title =
+        "${LocalizationUtil.localizedString(NetworkUtil.companyCodeKMB, Stores.localizationStore.localizationPref)}";
+
+    if (filterKeyword.isNotEmpty) {
+      title = "${title} (${displayedDirectionalRoutesForKMB.length})";
+    }
+
+    return title;
   }
 }
